+++ scripts/winesapos-tests.sh
@@ -1159,6 +1159,7 @@ echo 'Testing that the PipeWire audio library is installed complete.'
 echo 'Testing that support for all file systems is installed...'
 pacman_search_loop \
   apfsprogs-git \
+  bcachefs-dkms \
   bcachefs-tools \
   btrfs-progs \
   ceph-bin \