+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add documentation on how to resize a Ventoy image.
 - Allow installing COSMIC, GNOME, Hyprland, and/or Sway during the first-time setup.
 - Remove tmpfs mount on '/var/tmp'.
+- Change Bcachefs to be installed from the new DKMS package.
 
 ## 4.4.0 - The Rolling Update